requirements.txt
# LangChain Core
langchain-core

# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub


streamlit
langchain
langchain-community
sentence-transformers
chromadb
python-dotenv
openai
tqdm
PyPDF2

# Document processing (for PyPDFLoader and TextLoader)
pypdf>=3.17.0

